move SelectMenuOptionAttribute to its own file
